requirements.txt
appdirs==1.4.3
atomicwrites==1.2.1
attrs==18.2.0
black==18.9b0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
coverage==4.5.2
coveralls==1.5.1
docopt==0.6.2
Flask==1.0.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
marshmallow==2.16.3
more-itertools==4.3.0
pluggy==0.8.0
psycopg2==2.7.6.1
py==1.7.0
pytest==4.0.0
pytest-cov==2.6.0
requests==2.20.1
six==1.11.0
toml==0.10.0
urllib3==1.24.1
Werkzeug==0.14.1
